requirements.txt
streamlit
playwright
python-dotenv
requests
langchain-core
langchain-groq
mcp-use
asyncio 